Gui Package refactor
