+++ src/main/java/com/xray/client/gui/manage/GuiBlockListScrollable.java
@@ -1,5 +1,7 @@
-package com.xray.client.gui;
+package com.xray.client.gui.manage;
 
+import com.xray.client.gui.utils.GuiBase;
+import com.xray.client.gui.GuiSelectionScreen;
 import com.xray.common.XRay;
 import com.xray.common.reference.BlockItem;
 import net.minecraft.client.Minecraft;
